simplify Android
