--- experimental/examples/imageviewer/shared/src/commonMain/kotlin/example/imageviewer/view/Image.common.kt
@@ -1,7 +1,0 @@
-package example.imageviewer.view
-
-import androidx.compose.runtime.Composable
-import example.imageviewer.model.ContentState
-
-@Composable
-internal expect fun Image(content: ContentState)
