+++ experimental/examples/imageviewer/shared/src/desktopMain/kotlin/example/imageviewer/view/ScalableImage.desktop.kt
@@ -1,23 +1,19 @@
 package example.imageviewer.view
 
-import androidx.compose.foundation.*
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.graphics.asSkiaBitmap
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.toAwtImage
 import androidx.compose.ui.graphics.toComposeImageBitmap
-import androidx.compose.ui.input.key.Key
-import androidx.compose.ui.input.key.key
-import androidx.compose.ui.input.key.type
-import androidx.compose.ui.input.key.KeyEventType
-import androidx.compose.ui.input.key.onPreviewKeyEvent
+import androidx.compose.ui.input.key.*
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.DpSize
-import example.imageviewer.model.ContentState
 import example.imageviewer.style.DarkGray
 import example.imageviewer.utils.cropBitmapByBounds
 import example.imageviewer.utils.cropImage
@@ -26,7 +22,6 @@ import org.jetbrains.compose.resources.ExperimentalResourceApi
 import org.jetbrains.compose.resources.orEmpty
 import org.jetbrains.compose.resources.rememberImageBitmap
 import org.jetbrains.compose.resources.resource
-import org.jetbrains.skia.Bitmap
 import java.awt.Rectangle
 import java.awt.image.BufferedImage
 import java.io.ByteArrayOutputStream
@@ -34,23 +29,27 @@ import javax.imageio.ImageIO
 
 @OptIn(ExperimentalComposeUiApi::class, ExperimentalResourceApi::class)
 @Composable
-actual fun Image(content: ContentState) {
+actual fun ScalableImage(image: ImageBitmap, swipeNext: () -> Unit, swipePrevious: () -> Unit) {
     val drag = remember { DragHandler() }
     val size = LocalWindowSize.current
-    val onUpdate = remember { {
-        content.state.value = content.state.value.copy(
-            mainImage = org.jetbrains.skia.Image.makeFromEncoded(
-                toByteArray(
-                    cropBitmapByScale(
-                        content.state.value.origin!!.toAwtImage(), //todo npe
-                        size,
-                        content.state.value.scale,
-                        drag
-                    )
+    val onUpdate = remember {
+        {
+            if (image != null) {
+                content.state.value = content.state.value.copy(
+                    mainImage = org.jetbrains.skia.Image.makeFromEncoded(
+                        toByteArray(
+                            cropBitmapByScale(
+                                image.toAwtImage(),
+                                size,
+                                content.state.value.scale,
+                                drag
+                            )
+                        )
+                    ).toComposeImageBitmap()
                 )
-            ).toComposeImageBitmap()
-        )
-    } }
+            }
+        }
+    }
     val scaleHandler = remember { ScaleHandler(content.state) }
     Surface(
         color = DarkGray,
@@ -71,6 +70,7 @@ actual fun Image(content: ContentState) {
                                 Key.DirectionLeft -> {
                                     content.swipePrevious()
                                 }
+
                                 Key.DirectionRight -> {
                                     content.swipeNext()
                                 }
@@ -90,11 +90,7 @@ actual fun Image(content: ContentState) {
 }
 
 
-fun ContentState.updateMainImage() {
-
-}
-
-fun toByteArray(bitmap: BufferedImage) : ByteArray {
+fun toByteArray(bitmap: BufferedImage): ByteArray {
     val baos = ByteArrayOutputStream()
     ImageIO.write(bitmap, "png", baos)
     return baos.toByteArray()
@@ -117,4 +113,28 @@ fun cropBitmapByScale(
         bitmap,
         Rectangle(crop.x, crop.y, crop.width - crop.x, crop.height - crop.y)
     )
+}
+
+class ScaleHandler(
+    private val scaleState: MutableState<Float>,
+    private val maxFactor: Float = 5f,
+    private val minFactor: Float = 1f
+) {
+    fun reset() {
+        if (scaleState.value > minFactor) {
+            scaleState.value = minFactor
+        }
+    }
+
+    fun onScale(scaleFactor: Float): Float {
+        scaleState.value = scaleState.value + scaleFactor - 1f
+
+        if (maxFactor < scaleState.value) {
+            scaleState.value = maxFactor
+        }
+        if (minFactor > scaleState.value) {
+            scaleState.value = minFactor
+        }
+        return scaleFactor
+    }
 }
